issue_backend/requirements.txt
amqp==5.1.0
attr==0.3.1
attrs==21.2.0
backports.cached-property==1.0.1
cached-property==1.5.2
cffi==1.15.0
classic-app-layer==0.0.1
classic-aspects==0.0.1
classic-components==0.0.1
classic-http-api==0.0.2
classic-http-auth==0.2.1
classic-messaging==0.0.2
classic-messaging-kombu==0.0.3
classic-sql-storage==0.0.1
click==7.1.2
cryptography==36.0.2
falcon==3.0.1
greenlet==1.1.2
gunicorn==20.0.4
importlib-metadata==4.11.3
importlib-resources==5.4.0
kombu==5.1.0
Mako==1.2.0
MarkupSafe==2.1.1
pycparser==2.21
pydantic==1.8.2
psycopg2==2.9.2
PyJWT==2.0.1
python-json-logger==2.0.2
SQLAlchemy==1.4.32
typing_extensions==4.1.1
vine==5.0.0
zipp==3.7.0
